pyproject.toml
[tool.poetry]
name = "stegano"
version = "0.11.3"
description = "A pure Python Steganography module."
authors = [
    "Cédric Bonhomme <cedric@cedricbonhomme.org>"
]
license = "GPL-3.0-or-later"

readme = "README.md"

homepage = "https://github.com/cedricbonhomme/Stegano"
repository = "https://github.com/cedricbonhomme/Stegano"
documentation = "https://stegano.readthedocs.io"

keywords = ["steganography", "security", "stegano"]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Security",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)"
]

include = [
    "README.md",
    "COPYING",
    "CHANGELOG.md",
    "docs/**/*",
    "bin/*"
]

[tool.poetry.scripts]
stegano-lsb = "bin.lsb:main"
stegano-red = "bin.red:main"
stegano-steganalysis-parity = "bin.parity:main"
stegano-steganalysis-statistics = "bin.statistics:main"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pillow = ">=9.5,<11.0"
piexif = "^1.1.3"
crayons = "^0.4.0"
opencv-python = "^4.8.1.78"

[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
flake8 = "^6.0.0"
nose2 = "^0.14.0"
Sphinx = "^6.2.1"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry>=1.3.2"]
build-backend = "poetry.masonry.api"

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_errors = false
ignore_missing_imports = true
strict_optional = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unreachable = true

show_error_context = true
pretty = true

exclude = "build|dist|docs|stegano.egg-info"

[tool.isort]
profile = "black"
